Enforce restrict access on some pages from routeProvider.
Add access property to each route, specifying who can access this page
(programmer, administrator). When someone with a role inferior from the one
speicified on route, tries to access the page, it will be redirected to main
page.

For example when a user tries to access "/new-user" page (that access is
restricted only to administrator) he will be redirected to /